miApiJWT/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-multipart